+++ lyr-core/src/main/java/com/lyr/rule/glue/ScanGlueSessionActiveWithLongIdleTimeoutRuleExecution.java
@@ -6,9 +6,11 @@ import com.lyr.rule.RuleExecutionStrategy;
 import com.lyr.rule.glue.config.ScanGlueSessionActiveWithLongIdleTimeoutRuleConfig;
 import com.lyr.services.glue.GlueConnector;
 import java.util.List;
+import lombok.extern.slf4j.Slf4j;
 import software.amazon.awssdk.services.glue.model.Session;
 import software.amazon.awssdk.services.glue.model.SessionStatus;
 
+@Slf4j
 public class ScanGlueSessionActiveWithLongIdleTimeoutRuleExecution
         implements RuleExecutionStrategy<ScanGlueSessionActiveWithLongIdleTimeoutRuleConfig> {
 
@@ -17,11 +19,19 @@ public class ScanGlueSessionActiveWithLongIdleTimeoutRuleExecution
 
     @Override
     public ImmutableList<Finding> execute(final ScanGlueSessionActiveWithLongIdleTimeoutRuleConfig parameters) {
-        return GlueConnector.create().getSessionHistory().stream()
+        final var listOfSessions = GlueConnector.create().getSessionHistory();
+        final var findings = listOfSessions.stream()
                 .filter(session -> ACTIVE_STATUSES.contains(session.status())
                         && session.idleTimeout() > parameters.getMaxIdleTimeoutInMinutes())
                 .map(Session::id)
                 .map(Finding::byId)
                 .collect(ImmutableList.toImmutableList());
+
+        log.atInfo()
+                .addArgument(findings.size())
+                .addArgument(listOfSessions.size())
+                .log("Found {} session(s) with problems out of {} session(s).");
+
+        return findings;
     }
 }