+++ lyr-core/src/test/java/com/lyr/cli/runners/ScanAwsEnvironmentRunnerTest.java
@@ -1,13 +1,15 @@
 package com.lyr.cli.runners;
 
+import static com.lyr.TestUtil.DUMMY_STRING;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.reset;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
-import com.lyr.TestUtil;
+import com.lyr.config.RuleSetConfig;
 import com.lyr.config.Settings;
 import com.lyr.report.Reporter;
 import com.lyr.report.ReporterFactory;
@@ -15,6 +17,7 @@ import com.lyr.rule.Rule;
 import com.lyr.rule.RuleFactory;
 import com.lyr.util.RuleDefinition;
 import java.util.Optional;
+import java.util.Set;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -24,14 +27,17 @@ class ScanAwsEnvironmentRunnerTest {
 
     static MockedStatic<RuleFactory> mockedRuleFactory = mockStatic(RuleFactory.class);
     static MockedStatic<ReporterFactory> mockedReporterFactory = mockStatic(ReporterFactory.class);
+    static MockedStatic<RuleSetConfig> mockedRuleSetConfig = mockStatic(RuleSetConfig.class);
+    static RuleSetConfig mockedRuleSetConfigInstance = mock(RuleSetConfig.class);
     static Rule mockedRule = mock(Rule.class);
     static Reporter mockedReporter = mock(Reporter.class);
 
     @BeforeEach
     void beforeEach() {
-        reset(mockedRule, mockedReporter);
+        reset(mockedRule, mockedReporter, mockedRuleSetConfigInstance);
         mockedRuleFactory.reset();
         mockedReporterFactory.reset();
+        mockedRuleSetConfig.reset();
         Settings.setAppSettings(null);
     }
 
@@ -39,14 +45,21 @@ class ScanAwsEnvironmentRunnerTest {
     static void afterAll() {
         mockedRuleFactory.closeOnDemand();
         mockedReporterFactory.closeOnDemand();
+        mockedRuleSetConfig.closeOnDemand();
     }
 
     @Test
     void testThatGivenNoUserDefinedRuleSetDefaultRulesAreCreatedAndEvaluatedAndAReportIsCreated() {
         // Given
         Settings.setAppSettings(Settings.builder().build());
-        final var numberOfDefaultRules = 5;
+        final var numberOfDefaultRules = 2;
+        final var ruleSetFromConfig = Set.of(
+                RuleDefinition.SCAN_DYNAMODB_TABLE_IDLE,
+                RuleDefinition.SCAN_GLUE_SESSION_ACTIVE_WITH_LONG_IDLE_TIMEOUT);
+
         // When
+        mockedRuleSetConfig.when(RuleSetConfig::getRuleSetConfig).thenReturn(mockedRuleSetConfigInstance);
+        when(mockedRuleSetConfigInstance.getAllAvailableRuleDefinition()).thenReturn(ruleSetFromConfig);
         mockedRuleFactory.when(() -> RuleFactory.createRule(any())).thenReturn(mockedRule);
         mockedReporterFactory.when(() -> ReporterFactory.createReporter(any())).thenReturn(mockedReporter);
 
@@ -56,38 +69,39 @@ class ScanAwsEnvironmentRunnerTest {
         mockedRuleFactory.verify(() -> RuleFactory.createRule(RuleDefinition.SCAN_DYNAMODB_TABLE_IDLE), times(1));
         mockedRuleFactory.verify(
                 () -> RuleFactory.createRule(RuleDefinition.SCAN_GLUE_SESSION_ACTIVE_WITH_LONG_IDLE_TIMEOUT), times(1));
-        mockedRuleFactory.verify(
-                () -> RuleFactory.createRule(RuleDefinition.SCAN_GLUE_SESSION_ACTIVE_WITH_LONG_IDLE_TIMEOUT), times(1));
-        mockedRuleFactory.verify(
-                () -> RuleFactory.createRule(RuleDefinition.SCAN_CLOUDWATCH_LOG_GROUP_WITHOUT_RETENTION_POLICY),
-                times(1));
-        mockedRuleFactory.verify(
-                () -> RuleFactory.createRule(RuleDefinition.SCAN_LAMBDA_FUNCTION_WITH_DISALLOWED_ARCHITECTURE),
-                times(1));
         verify(mockedRule, times(numberOfDefaultRules)).evaluate();
         verify(mockedReporter, times(1)).report(any());
+        mockedRuleSetConfig.verify(RuleSetConfig::loadDefaultRuleSetConfig, times(1));
     }
 
     @Test
     void thatThatGivenUserDefinedRuleSetUserDefinedRulesAreCreatedAndEvaluatedAndAReportIsCreated() {
         // Given
-        final var testUserRuleSetAbsolutePath =
-                TestUtil.getAbsoluteFilePathOfResource("com/lyr/cli/runners/testUserConfig.yaml");
-        final var numberOfRulesInTestUserRuleSet = 1;
+        final var numberOfRulesInTestUserRuleSet = 2;
         final var settings = Settings.builder()
-                .userRuleSetConfigPath(Optional.of(testUserRuleSetAbsolutePath))
+                .userRuleSetConfigPath(Optional.of(DUMMY_STRING))
                 .build();
         Settings.setAppSettings(settings);
+        final var ruleSetFromConfig = Set.of(
+                RuleDefinition.SCAN_LAMBDA_FUNCTION_WITH_UNBOUNDED_CONCURRENCY,
+                RuleDefinition.SCAN_CLOUDWATCH_LOG_GROUP_WITHOUT_RETENTION_POLICY);
 
         // When
+        mockedRuleSetConfig.when(RuleSetConfig::getRuleSetConfig).thenReturn(mockedRuleSetConfigInstance);
+        when(mockedRuleSetConfigInstance.getAllAvailableRuleDefinition()).thenReturn(ruleSetFromConfig);
         mockedRuleFactory.when(() -> RuleFactory.createRule(any())).thenReturn(mockedRule);
         mockedReporterFactory.when(() -> ReporterFactory.createReporter(any())).thenReturn(mockedReporter);
 
         ScanAwsEnvironmentRunner.run();
 
         // Then
-        mockedRuleFactory.verify(() -> RuleFactory.createRule(RuleDefinition.SCAN_DYNAMODB_TABLE_IDLE), times(1));
+        mockedRuleFactory.verify(
+                () -> RuleFactory.createRule(RuleDefinition.SCAN_LAMBDA_FUNCTION_WITH_UNBOUNDED_CONCURRENCY), times(1));
+        mockedRuleFactory.verify(
+                () -> RuleFactory.createRule(RuleDefinition.SCAN_CLOUDWATCH_LOG_GROUP_WITHOUT_RETENTION_POLICY),
+                times(1));
         verify(mockedRule, times(numberOfRulesInTestUserRuleSet)).evaluate();
         verify(mockedReporter, times(1)).report(any());
+        mockedRuleSetConfig.verify(() -> RuleSetConfig.loadUserRuleSetConfig(DUMMY_STRING), times(1));
     }
 }